src/customskills/ocr/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
pydantic_core==2.14.5
httpx==0.25.2
python-dotenv==1.0.0
uvicorn==0.24.0.post1
requests==2.31.0
azure-identity
azure-search-documents==11.4.0
azure-storage-blob
bs4
langchain==0.1.0
langchain-community==0.0.9
msal>=0.6.1,<2
openai==1.6.1
azure-ai-documentintelligence
pandas
pypdf
azure-cognitiveservices-speech
PyPDF2
nest_asyncio
python-docx
python-dotenv
requests>=2,<3
tiktoken
spacy
azure-ai-formrecognizer
docx2txt
